# src/main/resources/swagger/swagger.yml
swagger: "2.0"
info:
  description: "This is a sample TODO application."
  version: "1.0.0"
  title: "Todo"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "jifardaniel@gmail.com"
host: "leanix.todos"
basePath: "/v1"
tags:
  - name: todo
schemes:
  - "https"
  - "http"
paths:
  /todos:
    get:
      tags:
        - "todo"
      summary: "Retrieve all todo list(s)"
      description: ""
      operationId: "getTodos"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Todo"
        "400":
          description: "Bad request"
        "500":
          description: "internal exception"
    post:
      tags:
        - "todo"
      summary: "Create new todo"
      description: ""
      operationId: "createTodo"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Todo object that needs to be added to the todo list"
          required: true
          schema:
            $ref: "#/definitions/Todo"
      responses:
        "201":
          description: "CREATED"
          schema:
            type: string
            example: "123"
        "400":
          description: "Bad request"
        "500":
          description: "internal exception"
  /todos/{id}:
    get:
      tags:
        - "todo"
      summary: "Retrieve a specific todo"
      description: "get a specific todo with id"
      operationId: "getTodo"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: "path"
          description: "ID of todo"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Todo"
        "400":
          description: "Bad request"
        "500":
          description: "internal exception"
    put:
      tags:
        - "todo"
      summary: "modify a specific todo"
      description: "update the object of a specific todo"
      operationId: "updateTodo"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: "path"
          description: "ID of todo"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Bad request"
        "500":
          description: "internal exception"
    delete:
      tags:
        - "todo"
      summary: "Remove a specific todo"
      description: "delete a spcific todo from a todos list"
      operationId: "deleteTodo"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: "path"
          description: "ID of todo"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Bad request"
        "500":
          description: "internal exception"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Todo:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      description:
        type: "string"
      tasks:
        $ref: "#/definitions/TodoTask"
    required:
      - id
      - name
  TodoTask:
    type: "object"
    properties:
      name:
        type: "string"
      description:
        type: "string"
    required:
      - name